Document Groups and Account tab
Documented groups and account tab. Added information about why groups and permissions may be used (sales territory example) as well as how to create a group, what can be edited within a group, and the need for a group for each entity, each attribute, multiple values.

Also added a small bit of information for the current Account tab.